Assignment4/Q2/requirements.txt
torch
torchvision
ftfy
regex
tqdm
pillow
numpy
pandas
matplotlib
git+https://github.com/openai/CLIP.git
